Safari-609.1.11.1/LayoutTests/fast/canvas/canvas-bezier-same-endpoint-expected.txt

Bug 105650: Test correct rendering of quadratic and bezier curves with coincident endpoints

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS blue_value is 255
PASS blue_value is 0
PASS blue_value is 255
PASS blue_value is 255
PASS blue_value is 0
PASS blue_value is 255
PASS blue_value is 0
PASS successfullyParsed is true

TEST COMPLETE
